+++ dbms/src/Functions/FunctionsExternalDictionaries.h
@@ -21,7 +21,7 @@
 #include <Interpreters/Context.h>
 #include <Interpreters/ExternalDictionariesLoader.h>
 
-#include <Functions/IFunction.h>
+#include <Functions/IFunctionImpl.h>
 #include <Functions/FunctionHelpers.h>
 
 #include <Dictionaries/FlatDictionary.h>