+++ dbms/src/Functions/base64Encode.cpp
@@ -1,5 +1,7 @@
 #include <Functions/FunctionFactory.h>
 #include <Functions/FunctionBase64Conversion.h>
+#include "config_functions.h"
+
 #if USE_BASE64
 #include <DataTypes/DataTypeString.h>