+++ dbms/src/IO/tests/gtest_bit_io.cpp
@@ -15,7 +15,7 @@
 #include <vector>
 #include <typeinfo>
 #include <iostream>
-
+#pragma GCC diagnostic ignored "-Wmissing-declarations"
 #include <gtest/gtest.h>
 
 using namespace DB;