+++ dbms/src/Interpreters/ExternalDictionariesLoader.cpp
@@ -1,7 +1,5 @@
 #include <Interpreters/ExternalDictionariesLoader.h>
-#include <Interpreters/Context.h>
 #include <Dictionaries/DictionaryFactory.h>
-#include <Dictionaries/getDictionaryConfigurationFromAST.h>
 
 namespace DB
 {
@@ -17,9 +15,13 @@ ExternalDictionariesLoader::ExternalDictionariesLoader(Context & context_)
 
 
 ExternalLoader::LoadablePtr ExternalDictionariesLoader::create(
-        const std::string & name, const Poco::Util::AbstractConfiguration & config, const std::string & key_in_config) const
+        const std::string & name, const Poco::Util::AbstractConfiguration & config,
+        const std::string & key_in_config, const std::string & repository_name) const
 {
-    return DictionaryFactory::instance().create(name, config, key_in_config, context);
+    /// For dictionaries from databases (created with DDL qureies) we have to perform
+    /// additional checks, so we identify them here.
+    bool dictionary_from_database = !repository_name.empty();
+    return DictionaryFactory::instance().create(name, config, key_in_config, context, dictionary_from_database);
 }
 
 void ExternalDictionariesLoader::addConfigRepository(
@@ -28,15 +30,4 @@ void ExternalDictionariesLoader::addConfigRepository(
     ExternalLoader::addConfigRepository(repository_name, std::move(config_repository), {"dictionary", "name"});
 }
 
-
-void ExternalDictionariesLoader::addDictionaryWithConfig(
-    const String & dictionary_name, const String & repo_name, const ASTCreateQuery & query, bool load_never_loading) const
-{
-    ExternalLoader::addObjectAndLoad(
-        dictionary_name, /// names are equal
-        dictionary_name,
-        repo_name,
-        getDictionaryConfigurationFromAST(query),
-        "dictionary", load_never_loading);
-}
 }