+++ docker/packager/binary/build.sh
@@ -13,7 +13,7 @@ cd build/build_docker
 ccache --show-stats ||:
 ccache --zero-stats ||:
 rm -f CMakeCache.txt
-cmake .. -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DSANITIZE=$SANITIZER $CMAKE_FLAGS
+cmake .. -LA -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DSANITIZE=$SANITIZER $CMAKE_FLAGS
 ninja
 ccache --show-stats ||:
 mv ./dbms/programs/clickhouse* /output
@@ -21,14 +21,14 @@ mv ./dbms/unit_tests_dbms /output
 find . -name '*.so' -print -exec mv '{}' /output \;
 find . -name '*.so.*' -print -exec mv '{}' /output \;
 
-count=`ls -1 /output/*.so 2>/dev/null | wc -l`
-if [ $count != 0 ]
+# May be set for split build or for performance test.
+if [ "" != "$COMBINED_OUTPUT" ]
 then
     mkdir -p /output/config
     cp ../dbms/programs/server/config.xml /output/config
     cp ../dbms/programs/server/users.xml /output/config
     cp -r ../dbms/programs/server/config.d /output/config
-    tar -czvf shared_build.tgz /output
+    tar -czvf "$COMBINED_OUTPUT.tgz" /output
     rm -r /output/*
-    mv shared_build.tgz /output
+    mv "$COMBINED_OUTPUT.tgz" /output
 fi